requirements.txt
aiogram
python-dotenv
matplotlib
openai
